requirements.txt
streamlit~=1.15.2
Pillow~=9.0.1
torch~=1.13.0
transformers~=4.25.1
sentencepiece~=0.1.97